Move `ty::tls` to seperate file
